refactor: decompose Timeline.tsx into focused sub-components
Extract 4 modules from Timeline.tsx (1442 -> 604 lines):
- TimelineWindowOverlay.tsx: window overlay component (107 lines)
- useTimelineScroll.ts: zoom, auto-scroll, wheel handler (285 lines)
- useTimelineDragSelection.ts: selection drag, window move/switch (306 lines)
- EmptyTrackRows.tsx: ArrangementEmptyTrackHeaderRow + EmptyTrackRow (236 lines)

All quality gates pass: tsc, tests (3008), build.

Closes #1027
